[all] Add stack-usage (4096) warning for each target
* add stack-usage warning for 4096 byte of stack;
* use space allocator for bit stack data;
* use emplace container api to reduce stack usage.

+++ include/aos/common/types.hpp
@@ -732,7 +732,7 @@ struct Host {
  */
 struct DeviceInfo {
     StaticString<cDeviceNameLen>                                  mName;
-    int                                                           mSharedCount {0};
+    size_t                                                        mSharedCount {0};
     StaticArray<StaticString<cGroupNameLen>, cMaxNumGroups>       mGroups;
     StaticArray<StaticString<cDeviceNameLen>, cMaxNumHostDevices> mHostDevices;
